Experiment with another xmobar setup
I was curious to try a contrast-y color rather than a blendy-inny color,
but then it looked a little awkward against the tmux status line. So I'm
trying out a sandwich where it has the same color as the tmux status
line, and it's at the top.

Needed to futz with some of the text colors to make them show up nicely
against the green background