Update desktop_xrandr for multi-monitor setup
The idea is that it can be nice to have the laptop open and to the left
of desktop monitor. I've been pulling up my calendar and putting it on
the smaller display to the side, available for glancing at as needed.

I could revert this in the future if I grow sick of this setup.