Switch to taffybar from xmobar
I mainly wanted it to have a system tray built in.

This had some pains:

needed to re-add the xmonad build script

needed to switch from pacman-installed xmonad and xmonad-contrib to
stack installed xmonad and xmonad-contrib. This is because taffybar is
available via stack, but not via pacman, and taffybar depends on xmonad.
I think that's true anyway.

So the pre-reqs for this are:

```
sudo pacman -Sy stack
stack install xmonad xmonad-contrib
stack install gtk2hs-buildtools
stack install taffybar
```

Also switching to a single-monitor setup even when plugged into a
desktop monitor, because taffybar 0.4.6 doesn't seem to support multiple
monitors. 0.5.0 does. That's fine though.

Much of this config was copied from either the taffybar examples or
wfleming's dotfiles.