Really switch from xmonad to i3
Follow up to the earlier one when I was just trying it. I've been pretty
happily using it. I'm not really a power user and don't find the tiling
behavior as intuitive but it feels more easy to maintain and my config
doesn't occasionally stop working when I update one of the haskell
packages that stream steadily thru the package manager