Switch to my-fancy-zsh-git-prompt
Must be built from source from
https://github.com/maxjacobson/my-fancy-zsh-git-prompt/

Behaves similarly to the old shell script but shows the current working
directory relative to the current git repository. It also loses the
dirty state tracking, because I couldn't quite get that working in Rust.
But overall I think it's a pretty good improvement.