Add clone helper script
I don't know if I actually will remember to use this, but it was a fun
little scripting challenge.

Probably it would be nicer if it were fully in Ruby, but I don't know
how to cd into a directory from a ruby script such that when the script
finishes, my CWD has actually changed.