Configure ale to use rubocop
This lets me see rubocop issues inline while editing.

This is most valuable if the codebase has all of its issues already
resolved, or deferred to .rubocop_todo.yml. Otherwise, you'll see loads
of issues all the time.

The thing with "bundle" makes sure that it runs the version of rubocop
specified in the project's Gemfile.