Some changes to clone helper
This makes it simpler in some ways and more complicated in other ways,
but I think it's kind of neat.

There's _multiple_ execs to achieve what I want...

- a function that changes directory
- complex logic in ruby, not shell
- when it's done, there's still only one process running

To achieve that, the shell function execs the ruby script, and the ruby
script execs a new shell. So when you use the function, you run a ruby
script (which starts a child process), which then replaces itself with a
shell process, and then the parent shell process replaces itself with
the new shell process.

Did that make sense?

Is this wise?

Are there better ways to do this?

Not sure!