Improve t tmux alias
This resolves a longstanding issue, which is that two different
directories which happen to share a basename will attach to the same
tmux session. This fix (which chatgpt wrote for me) is to generate a
hash of the absolute path, and include that in the session name. So
~/src/gh/maxjacobson/seasoning becomes "seasoning 3f2c" but
~/Desktop/seasoning becomes "seasoning 8870". That's, admittedly, a
little less visually appealing than just "seasoning", but I like that
it's still succinct while being unambiguous.