Set up homebrew and ~/bin in conf.d script
I determined that my earlier setup was working but only because of fish
universal variables. I must have used `fish_add_path /opt/homebrew/bin`
at some point and it then remembered indefinitely that this is on my
path. I don't think I actually want that. With this approach, no path
manipulation is sticky, it must be set on every shell initialization.
And it must be set early enough in the initialization flow that it will
be set before things that depend on it being set run.