Update fish prompt
This adds an emoji to indicate that a fetch is in progress, which will
also make it more visible when a fetch has failed and we're in a stuck
state where we think it's fetching even though it's not. That should get
auto-cleaned up, but now it should at least be more visible.

This also changes from fish_git_prompt to fish_vcs_prompt in case I
actually do start using jj in some repos, it will have a jj prompt
instead of a git prompt. So far the jj prompt seems kind of broken.
Nothing is displayed at all. Maybe I'm supposed to like that.